converted meaning page to dialog box
